Show number of users who enabled auto-enroll option
Enhancement on functionality that shows the number of users that has
enabled some feature. Now the option "Automatically enable all new beta
features" also shows the number of users that have enabled it, like the
other features in the same list. For this, I created the
getBetaPreferences method in the BetaFeaturesHooks class, which will make the
auto enroll option handled like the others. It was also necessary to modify the
HTMLFeatureField class methods in order to make them able to handle
auto-enroll fields.

Bug: 59955
Change-Id: Ie041a553d062ebf5c674a287bca8c42e203a48d3